Pull apart generic vs Liveblocks-specific "utils"
