+++ packages/liveblocks-core/src/immutable.ts
@@ -6,12 +6,8 @@ import { LiveObject } from "./LiveObject";
 import { LiveRegister } from "./LiveRegister";
 import type { LiveNode, Lson, LsonObject, ToJson } from "./types/Lson";
 import type { StorageUpdate } from "./types/StorageUpdates";
-import {
-  findNonSerializableValue,
-  isLiveList,
-  isLiveObject,
-  isPlainObject,
-} from "./utils";
+import { findNonSerializableValue, isLiveList, isLiveObject } from "./utils";
+import { isPlainObject } from "./lib/utils";
 
 function lsonObjectToJson<O extends LsonObject>(
   obj: O
